feat(email): added email notifications when content is shared
