--- web/Pages/API/Index.cshtml
@@ -1,5 +1,0 @@
-@page
-@model Atlas_Web.Pages.API.IndexModel
-@{
-    ViewData["Title"] = "Index";
-}
