--- web/Pages/API/Partials/_EmailTemplate.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html
-  xmlns="http://www.w3.org/1999/xhtml"
-  xmlns:v="urn:schemas-microsoft-com:vml"
-  xmlns:o="urn:schemas-microsoft-com:office:office"
-  lang="en"
->
-  <head>
-    <title>Message from Atlas BI Library</title>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <style type="text/css">
-      body { margin:0;padding:0;-webkit-text-size-adjust:100%;-ms-text-size-adjust:100%;background-color:#FAFAFAFF;font-family: &apos;book antiqua&apos;, palatino, serif; font-size: 16px; color: #404040; }
-      p { margin: 0px; }
-      .box {margin-top:30px;}
-    </style>
-  </head>
-  <body>
-    <div>
-      <div class="box" style="text-align: center; max-width: 600px">
-        <p style="font-size: 36px">Atlas</p>
-        <p><em style="font-size: 16px">of</em></p>
-        <p style="font-size: 24px">Information Management</p>
-      </div>
-      <div class="box" style="">
-        <div>
-          <p>Hi {ToUserFirstname},</p>
-          <br />
-          <p>{myMessage}</p>
-          <br />
-          <p><a href="{url}" target="_blank">{Link}</a></p>
-          <br />
-          <p>Regards,</p>
-          <p>{FromUserName}</p>
-          <p>{FromUserEmail}</p>
-        </div>
-      </div>
-      <div
-        class="box"
-        style="
-          font-family: Ubuntu, Helvetica, Arial, sans-serif;
-          font-size: 11px;
-          line-height: 1.5;
-          color: #000000;
-        "
-      >
-        <p>
-          <span style="color: #34495e"
-            >This email automatically sent from
-            <a
-              style="color: #34495e"
-              title="Atlas of Information Management"
-              href="https://Atlas"
-              target="_blank"
-              rel="noopener"
-              >Atlas</a
-            >, Riverside Healthcare's official Report Library</span
-          >
-        </p>
-      </div>
-    </div>
-  </body>
-</html>
