--- web/Pages/API/Partials/_Html.cshtml
@@ -1,1 +1,0 @@
-@Html.Raw(ViewBag.Html)
