+++ web/Pages/Reports/_Modals.cshtml
@@ -65,16 +65,16 @@
             <div class="box">
                 <strong class="mb-4">Share @Model.Name</strong>
 
-                <form class="mlbx-newMsgForm" action="/mail?handler=SendMail" method="post">
+                <form class="share" action="/mail?handler=SendMail" method="post">
                     <input type="hidden" class="mlbx-share" value="1" />
-                    <input type="hidden" class="mlbx-shareName" value="@Model.Name" />
-                    <input type="hidden" class="mlbx-shareUrl" value="@Context.Request.HttpContext.Request.Scheme://@Context.Request.HttpContext.Request.Host/reports?id=@Model.ReportObjectId" />
+                    <input type="hidden" class="share-name" value="@Model.Name" />
+                    <input type="hidden" class="share-url" value="@Context.Request.HttpContext.Request.Scheme://@Context.Request.HttpContext.Request.Host/reports?id=@Model.ReportObjectId" />
 
                     <input type="hidden" id="convid" />
 
                     <div class="field mt-4">
                         <label class="label">To:</label>
-                        <div class="mini-tags field is-grouped is-grouped-multiline mail-to" required name="to"></div>
+                        <div class="mini-tags field is-grouped is-grouped-multiline share-to" required name="to"></div>
                         <div class="field">
                             <div class="control is-relative is-expanded">
                                 <input class="input input-mini multiselect mini-close-fast" placeholder="search for someone.." search-area="UserSearch" />
@@ -91,12 +91,12 @@
                     <div class="field">
                         <label class="label">Subject</label>
                         <div class="control">
-                            <input class="input mail-subject" placeholder="e.g hey, check this out" value="[Share] @Model.Name" />
+                            <input class="input share-subject" placeholder="e.g hey, check this out" value="[Share] @Model.Name" />
                         </div>
                     </div>
                     <div class="field">
                         <label class="label">Message</label>
-                        <div class="editor" data-inputName="mail-message">
+                        <div class="editor" data-inputName="share-message">
                             <textarea class="textarea">
 Hi!
