+++ web/Pages/Shared/_Messagebox.cshtml
@@ -1,10 +1,1 @@
-<div class="message-wrapper hidden">
-	<div class="message-wrapper-inner">
-    	<button class="message-close"></button>
-    	<div class="message-container">
-    		<div class="message-inner">
-    			nice message here.
-    		</div>
-    	</div>
-    </div>
-</div>
+<div class="fixed-notification-wrapper"></div>
