+++ web/Pages/Users/Sections/_Settings.cshtml
@@ -1,0 +1,12 @@
+<h3 class="title is-3">Settings</h3>
+
+<h4 class="title is-4">Email</h4>
+
+<div class="field toggle">
+        <input type="hidden"  value="">
+        <input id="switch-hidden" type="checkbox" name="switch-hidden" class="switch is-rounded is-info"
+        @if(Model.Initiative.Hidden == "Y"){
+            <text>checked="checked"</text>
+        } >
+        <label for="switch-hidden">Enable Share Notification</label>
+    </div>
