+++ web/Pages/Users/_Links.cshtml
@@ -31,6 +31,12 @@
                 <a class="tab-lnk" href="#analytics">Analytics</a>
             </li>
         }
+        @if (@Model.UserId == @Model.MyId)
+        {
+            <li>
+                <a class="tab-lnk" href="#settings">Settings</a>
+            </li>
+        }
     </ul>
 </nav>
