+++ web/Services/IRazorPartialToStringRenderer.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+
+using Microsoft.AspNetCore.Mvc.ViewFeatures;
+
+namespace Atlas_Web.Services
+{
+    public interface IRazorPartialToStringRenderer
+    {
+        Task<string> RenderPartialToStringAsync(string partialName, ViewDataDictionary ViewData);
+    }
+}
