+++ web/appsettings.json
@@ -1,7 +1,7 @@
 {
   "AllowedHosts": "*",
   "AppSettings": {
-    "default_from_email_address": "cool@example.net",
+    "base_url": "https://atlas.example.org",
     "manage_engine_server": "https://example.net:8080",
     "manage_engine_tech_key": "123456789",
     "org_ad_domain": "YOURORGAD",
@@ -9,7 +9,10 @@
     "org_name": "Example Healthcare",
     "smtp_port": 465,
     "smtp_server": "mail.example.net",
-    "smtp_use_ssl": true
+    "smtp_use_ssl": true,
+    "smtp_sender_name": "Atlas | Example Healthcare Analytics",
+    "smtp_sender_email": "atlas@example.com",
+    "smtp_default_reply_email": "atlas@example.com"
   },
   "ConnectionStrings": {
     "AtlasDatabase": "Server=127.0.0.1;Database=atlas;User Id=SA; Password=p@ssw0rd; MultipleActiveResultSets=true"
