+++ web/wwwroot/css/messagebox.scss
@@ -1,65 +1,9 @@
-.message-wrapper {
-  background-color: #fff;
-  border: 1px solid #b7b7b7;
-  border-radius: 6px;
-  border-top: 1px solid #d0d0d0;
-  box-shadow: 0 0 30px 10px #d1d1d1;
-  font-size: 14px;
+.fixed-notification-wrapper {
   height: auto;
-  left: calc(50% - 350px / 2);
-  margin-top: 10px;
-  opacity: 1;
+  right: 20px;
   position: fixed;
-  top: 0;
+  top: 20px;
   transition: all 0.5s;
-  width: 350px;
+  width: auto;
   z-index: 999;
-}
-
-.message-wrapper-inner {
-  border-radius: inherit;
-  border-top: 1px solid #f4f5f6;
-}
-
-.message-wrapper.hidden {
-  opacity: 0;
-  top: -200px;
-  transition: all 1s;
-}
-
-.message-wrapper button {
-  border-radius: 6px;
-  height: 12px;
-  position: absolute;
-  top: 6px;
-  transition: all 0.2s;
-  width: 12px;
-}
-
-.message-wrapper .message-close {
-  background-color: #ff564f;
-  border: 1px solid #dc413a;
-  left: 8px;
-}
-
-.message-container {
-  border-radius: inherit;
-  border-top: 22px solid #e6e3e5;
-}
-
-.message-container::before {
-  color: #393939;
-  content: 'Message';
-  font-size: 13px;
-  left: 50%;
-  letter-spacing: 0.4px;
-  margin-left: -27px;
-  position: absolute;
-  top: 4px;
-  width: 54px;
-}
-
-.message-container .message-inner {
-  border-top: 1px solid #cccacc;
-  padding: 8px;
 }
